# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/tenants/{tenantId}/flags": {
      "post": {
        "description": "Creates a flag and all three environment configurations at once. Environments omitted from `environments` default to disabled at 0% rollout, so a new flag is inert until explicitly turned on.\n\n`flagKey` is unique per tenant rather than globally, so two tenants may each own a flag called `new-checkout`.",
        "operationId": "FlagsController_create",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "description": "Tenant that owns the flag. Must match the tenant behind the API key.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFlagDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Flag created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureFlagDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (unknown flag type, variant weight out of range, …).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `x-api-key`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The API key belongs to a different tenant than the `tenantId` in the path.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "A flag with this key already exists for the tenant.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Create a feature flag",
        "tags": [
          "Feature Flags"
        ]
      },
      "get": {
        "description": "Returns the tenant's flags. Filtering by `environment` narrows each flag's `environments` array to just that environment, which is the shape a single-environment dashboard wants.",
        "operationId": "FlagsController_findAll",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "description": "Tenant that owns the flag. Must match the tenant behind the API key.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "environment",
            "required": false,
            "in": "query",
            "description": "Return only this environment's configuration for each flag. Omit for all three.",
            "schema": {
              "enum": [
                "development",
                "staging",
                "production"
              ],
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by archive state. Omit to return both active and archived flags.",
            "schema": {
              "enum": [
                "active",
                "archived"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching flags.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeatureFlagDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `x-api-key`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The API key belongs to a different tenant than the `tenantId` in the path.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "List feature flags",
        "tags": [
          "Feature Flags"
        ]
      }
    },
    "/api/v1/tenants/{tenantId}/flags/{flagKey}": {
      "put": {
        "description": "Partial update — send only the fields that change. Toggle, rollout percentage, targeting and variants are per-environment and require `environment` in the body; name, description and default value apply across all environments.\n\nEvery change is written to the append-only audit log with the full before and after state, and invalidates the tenant's evaluation cache so it takes effect on the very next evaluation.",
        "operationId": "FlagsController_update",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "description": "Tenant that owns the flag. Must match the tenant behind the API key.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "flagKey",
            "required": true,
            "in": "path",
            "description": "Flag key.",
            "schema": {
              "example": "new-checkout-flow",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFlagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated flag.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureFlagDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed — for example an unrecognised `environment` value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `x-api-key`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The API key belongs to a different tenant than the `tenantId` in the path.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "No such flag for this tenant.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Update a flag",
        "tags": [
          "Feature Flags"
        ]
      },
      "delete": {
        "description": "Sets `isArchived`; the row is never removed. Archived flags stop being evaluated and drop out of `?status=active`, but stay readable via `?status=archived` and keep their full audit history — hard-deleting would destroy the compliance trail the audit log exists to provide.",
        "operationId": "FlagsController_archive",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "description": "Tenant that owns the flag. Must match the tenant behind the API key.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "flagKey",
            "required": true,
            "in": "path",
            "description": "Flag key.",
            "schema": {
              "example": "legacy-banner",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Flag archived. No response body."
          },
          "401": {
            "description": "Missing or invalid `x-api-key`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The API key belongs to a different tenant than the `tenantId` in the path.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "No such flag for this tenant.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Archive a flag (soft delete)",
        "tags": [
          "Feature Flags"
        ]
      }
    },
...
  },
  "info": {
    "title": "Multi-Tenant Feature Flag Service",
...
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local (Docker Compose)"
    },
    {
      "url": "https://feature-flag-service-staging-dhqwxiey3q-uc.a.run.app",
      "description": "Staging (Cloud Run)"
    },
    {
      "url": "https://feature-flag-service-production-<hash>-uc.a.run.app",
      "description": "Production (Cloud Run)"
    }
  ],
  "components": {
    "securitySchemes": {
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Tenant API key, returned once at registration. Identifies the tenant for every scoped operation."
      },
      "admin-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-key",
        "description": "Platform admin key, required for tenant registration wherever ADMIN_API_KEY is configured. Read it with: gcloud secrets versions access latest --secret=feature-flag-service-<env>-admin-api-key"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "description": "Either a single message or, for validation failures, one entry per invalid field.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": [
              "environment must be one of the following values: development, staging, production"
            ]
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "FlagType": {
        "type": "string",
        "enum": [
          "boolean",
          "string",
          "number"
        ],
        "description": "Value type this flag resolves to."
      },
      "Environment": {
        "type": "string",
        "enum": [
          "development",
          "staging",
          "production"
        ],
        "description": "Environment this configuration applies to."
      },
      "VariantDto": {
        "type": "object",
        "properties": {
          "value": {
            "description": "Value served when this variant is selected. Should match the flag type.",
            "example": "treatment",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              }
            ]
          },
          "weight": {
            "type": "number",
            "description": "Share of the bucket space for this variant. Weights across all variants should total 100.",
            "example": 50,
            "minimum": 0,
            "maximum": 100
          }
        },
        "required": [
          "value",
          "weight"
        ]
      },
      "EnvironmentConfigDto": {
        "type": "object",
        "properties": {
          "environment": {
            "description": "Environment this configuration applies to.",
            "example": "production",
            "allOf": [
              {
                "$ref": "#/components/schemas/Environment"
              }
            ]
          },
          "isEnabled": {
            "type": "boolean",
            "description": "Master on/off switch for this environment. When false the flag always resolves to defaultValue.",
            "example": true,
            "default": false
          },
          "rolloutPercentage": {
            "type": "number",
            "description": "Percentage of users receiving the enabled value. Membership is deterministic per user — see the evaluation algorithm in the README.",
            "example": 25,
            "minimum": 0,
            "maximum": 100,
            "default": 0
          },
          "targeting": {
            "type": "object",
            "description": "Reserved for attribute-based targeting rules. Accepted and persisted, not yet consulted during evaluation.",
            "additionalProperties": true,
            "example": {}
          },
          "variants": {
            "description": "Weighted variants for A/B testing. String flags only. Variant selection is independent of rollout membership.",
            "example": [
              {
                "value": "control",
                "weight": 50
              },
              {
                "value": "treatment",
                "weight": 50
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VariantDto"
            }
          }
        },
        "required": [
          "environment"
        ]
      },
      "CreateFlagDto": {
        "type": "object",
        "properties": {
          "flagKey": {
            "type": "string",
            "description": "Tenant-scoped identifier used in evaluation calls and URLs. Unique per tenant, not globally.",
            "example": "new-checkout-flow"
          },
          "name": {
            "type": "string",
            "description": "Human-readable flag name.",
            "example": "New Checkout Flow"
          },
          "description": {
            "type": "string",
            "description": "Free-text description of what the flag controls.",
            "example": "Routes traffic to the rebuilt checkout funnel."
          },
          "type": {
            "description": "Value type this flag resolves to.",
            "example": "boolean",
            "allOf": [
              {
                "$ref": "#/components/schemas/FlagType"
              }
            ]
          },
          "defaultValue": {
            "description": "Value returned when the flag is disabled or the user falls outside the rollout. Should match `type`.",
            "example": false,
            "oneOf": [
              {
                "type": "boolean"
              },
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ]
          },
          "environments": {
            "description": "Per-environment configuration. All three environments are always created; any omitted here default to disabled at 0% rollout.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EnvironmentConfigDto"
            }
          }
        },
        "required": [
          "flagKey",
          "name",
          "type",
          "defaultValue"
        ]
      },
      "FlagEnvironmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "flagId": {
            "type": "string",
            "format": "uuid",
            "description": "The flag this configuration belongs to."
          },
          "environment": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Environment"
              }
            ]
          },
          "isEnabled": {
            "type": "boolean",
            "example": true
          },
          "rolloutPercentage": {
            "type": "number",
            "example": 25,
            "minimum": 0,
            "maximum": 100
          },
          "targeting": {
            "type": "object",
            "additionalProperties": true,
            "example": {}
          },
          "variants": {
            "nullable": true,
            "description": "Null when the flag has no variants configured.",
            "example": [
              {
                "value": "control",
                "weight": 50
              },
              {
                "value": "treatment",
                "weight": 50
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VariantDto"
            }
          }
        },
        "required": [
          "id",
          "flagId",
          "environment",
          "isEnabled",
          "rolloutPercentage",
          "targeting",
          "variants"
        ]
      },
      "FeatureFlagDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "flagKey": {
            "type": "string",
            "example": "new-checkout-flow"
          },
          "name": {
            "type": "string",
            "example": "New Checkout Flow"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/FlagType"
              }
            ]
          },
          "defaultValue": {
            "description": "Value returned when disabled or outside the rollout.",
            "example": false,
            "oneOf": [
              {
                "type": "boolean"
              },
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ]
          },
          "isArchived": {
            "type": "boolean",
            "description": "Soft-delete marker. Archived flags are excluded from evaluation.",
            "example": false
          },
          "environments": {
            "description": "One entry per environment. Filtered to a single entry when the list endpoint is called with `?environment=`.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FlagEnvironmentDto"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "tenantId",
          "flagKey",
          "name",
          "type",
          "defaultValue",
          "isArchived",
          "environments",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateFlagDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "New human-readable name. Applies across all environments.",
            "example": "New Checkout Flow (v2)"
          },
          "description": {
            "type": "string",
            "description": "New description. Applies across all environments.",
            "example": "Rebuilt checkout funnel, now including express pay."
          },
          "defaultValue": {
            "description": "New fallback value for disabled / not-in-rollout evaluations. Applies across all environments.",
            "example": false,
            "oneOf": [
              {
                "type": "boolean"
              },
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ]
          },
          "environment": {
            "description": "Environment to modify. REQUIRED for any of the per-environment fields below to take effect.",
            "example": "production",
            "allOf": [
              {
                "$ref": "#/components/schemas/Environment"
              }
            ]
          },
          "isEnabled": {
            "type": "boolean",
            "description": "Toggle the flag on or off in the given `environment`. Requires `environment`.",
            "example": true
          },
          "rolloutPercentage": {
            "type": "number",
            "description": "New rollout percentage in the given `environment`. Requires `environment`. Users already inside the rollout remain inside when this increases.",
            "example": 50,
            "minimum": 0,
            "maximum": 100
          },
          "targeting": {
            "type": "object",
            "description": "Targeting rules for the given `environment`. Requires `environment`.",
            "additionalProperties": true,
            "example": {}
          },
          "variants": {
            "description": "Replacement variant set for the given `environment`. Requires `environment`. Re-weighting does not reshuffle who is inside the rollout.",
            "example": [
              {
                "value": "control",
                "weight": 30
              },
              {
                "value": "treatment",
                "weight": 70
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VariantDto"
            }
          }
        }
      },
...
    }
  }
}